# Pertemuan12/tugas/backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TES SWAGGER PEMROGRAMAN III",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/indrariksa",
            "email": "indra@ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Melakukan proses login dan mengembalikan token PASETO jika username dan password valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "Login credentials (username dan password)",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login success",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid body"
                    },
                    "401": {
                        "description": "Username not found or Wrong password"
                    },
                    "500": {
                        "description": "Failed to generate token"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Membuat akun baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "description": "Register credentials (username, password, and role)",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register success",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid body"
                    },
                    "401": {
                        "description": "Username, password, and role are required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Login success"
                },
                "token": {
                    "type": "string",
                    "example": "v4.public.xxxxxx"
                }
            }
        },
...
        "model.RegisterResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "123456abcdef"
                },
                "message": {
                    "type": "string",
                    "example": "Register success"
                }
            }
        },
        "model.UserLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}